AWS Batch Worker/requirements.txt
# --- GPU HARDWARE LAYER ---
# Explicitly pinned to match CUDA 12.1 in the base image
torch==2.5.1
torchvision==0.20.1
onnxruntime-gpu==1.19.2

# --- YOUR LOCAL VERSIONS ---
unstructured==0.18.27
langchain_core==1.2.7
langchain_google_genai==4.2.0
langchain_groq==1.1.1
boto3==1.42.32
chromadb==1.4.1
SQLAlchemy==2.0.46
python-dotenv==1.2.1

# --- DATABASE DRIVER ---
psycopg2-binary